Fixed some more bad practices. Fixed some infinite recursion errors. Fixed some errors when calculating graph connectivity. Removed unused set in Krusal algorithm.

+++ src/com/syncleus/dann/math/visualization/MathFunctionDataBinder.java
@@ -84,7 +84,7 @@ public final class MathFunctionDataBinder implements Binned2DData
 		if (newZMax == newZMin)
 		{
 			newZMax += (float) 1.0;
-			newZMin -= 1.0;
+			newZMin -= (float) 1.0;
 		}
 		this.maxZ = newZMax;
 		this.minZ = newZMin;
